Improved some types; Improved database result matching

+++ models/core/BetSpread.py
@@ -1,10 +1,10 @@
 from pydantic import BaseModel
 
 class BetSpread(BaseModel):
-  true_zero: int
-  true_one: int
-  true_two: int
-  true_three: int
-  true_four: int
-  true_five: int
-  true_six: int
+  true_zero: float
+  true_one: float
+  true_two: float
+  true_three: float
+  true_four: float
+  true_five: float
+  true_six: float
